refactor: 共通フィクスチャをconftest.pyに集約しテスト間の重複を排除
テストガイドラインの「共通fixtureはconftest.pyに配置」に従い、
各テストファイルに散在していたフィクスチャをconftest.pyに昇格。
ハンズオンドキュメント(04, 05)もconftest.pyパターンに更新。
