+++ tests/unit/services/hotel/test_reserve_hotel.py
@@ -1,5 +1,4 @@
 from decimal import Decimal
-from unittest.mock import MagicMock
 
 from services.hotel.applications.reserve_hotel import ReserveHotelService
 from services.hotel.domain.entity.hotel_booking import HotelBooking
@@ -8,15 +7,12 @@ from services.hotel.domain.factory.hotel_booking_factory import (
     HotelBookingFactory,
     HotelDetails,
 )
-from services.shared.domain import TripId
 
 
 class TestReserveHotelService:
-    def test_reserve_creates_and_saves_booking(self):
-        mock_repository = MagicMock()
+    def test_reserve_creates_and_saves_booking(self, mock_repository, trip_id):
         factory = HotelBookingFactory()
         service = ReserveHotelService(repository=mock_repository, factory=factory)
-        trip_id = TripId(value="trip-123")
         hotel_details: HotelDetails = {
             "hotel_name": "Grand Hotel",
             "check_in_date": "2024-01-01",
