+++ tests/unit/services/payment/test_process_payment.py
@@ -1,19 +1,17 @@
 from decimal import Decimal
-from unittest.mock import MagicMock
 
 from services.payment.applications.process_payment import ProcessPaymentService
 from services.payment.domain.entity.payment import Payment
 from services.payment.domain.enum.payment_status import PaymentStatus
 from services.payment.domain.factory.payment_factory import PaymentFactory
-from services.shared.domain import TripId
 
 
 class TestProcessPaymentService:
-    def test_process_creates_completes_and_saves_payment(self):
-        mock_repository = MagicMock()
+    def test_process_creates_completes_and_saves_payment(
+        self, mock_repository, trip_id
+    ):
         factory = PaymentFactory()
         service = ProcessPaymentService(repository=mock_repository, factory=factory)
-        trip_id = TripId(value="trip-123")
 
         payment = service.process(
             trip_id=trip_id,
